refactor(kubectl): split the package into one module per domain
@zuke/kubectl was a single 1,371-line kubectl.ts holding the settings
base, seventeen command classes, the JSON narrowing helpers and the task
object. That is against the one-domain-per-file guideline, and it is why
each addition to the package made the file worse rather than better.

Move each command's settings into the module for its domain — settings,
manifests, resources, workloads, pods, diagnostics, and the resource JSON
parsing — and leave kubectl.ts as the assembler that composes the task
object from them. mod.ts re-exports every module, so the public surface is
unchanged; the existing tests pass untouched, which is what makes this a
move rather than a rewrite.
